+++ src/components/pages/TokenMarket/signals.ts
@@ -1,4 +1,4 @@
-"use client"
+"use client";
 
 import { Listing } from "@/types/bsv20";
 import { BSV20TXO } from "@/types/ordinals";
@@ -6,12 +6,13 @@ import { Signal } from "@preact/signals-react";
 
 export const listings = new Signal<Listing[] | null>(null);
 export const sales = new Signal<BSV20TXO[] | null>(null);
+export const myListings = new Signal<Listing[] | null>(null);
+export const mySales = new Signal<BSV20TXO[] | null>(null);
 
 listings.subscribe((listings) => {
-    // localStorage.setItem("listings", JSON.stringify(listings));
-})
+	// localStorage.setItem("listings", JSON.stringify(listings));
+});
 
 sales.subscribe((sales) => {
-  
-    // localStorage.setItem("sales", JSON.stringify(sales));
-})
+	// localStorage.setItem("sales", JSON.stringify(sales));
+});
